requirements.txt
streamlit
mediapipe==0.10.9
pillow
scikit-learn
numpy
opencv-python-headless
